Bug 390713 - Tab traversal broken in 64-bit ToolBar

+++ bundles/org.eclipse.swt/Eclipse SWT/cocoa/org/eclipse/swt/widgets/ToolBar.java
@@ -476,7 +476,7 @@ public int getRowCount () {
 	return layout (rect.width, rect.height, false) [0];
 }
 
-boolean hasKeyboardFocus(int inId) {
+boolean hasKeyboardFocus(long /*int*/ inId) {
 	return hasFocus();
 }
